[ADF-2065] renamed components to keep backward compatibility
