+++ lib/content-services/content-node-selector/content-node-selector-panel.component.scss
@@ -1,0 +1,161 @@
+@mixin adf-content-node-selector-theme($theme) {
+    $primary: map-get($theme, primary);
+    $foreground: map-get($theme, foreground);
+    $background: map-get($theme, background);
+
+    .adf-content-node-selector {
+
+        &-content {
+            padding-top: 0;
+
+            &-input {
+                width: 100%;
+
+                &-icon {
+                    color: mat-color($foreground, disabled-button);
+                    cursor: pointer;
+
+                    &:hover {
+                        color: mat-color($foreground, base);
+                    }
+                }
+            }
+
+            .mat-input-underline .mat-input-ripple {
+                height: 1px;
+                transition: none;
+            }
+
+            .adf-site-dropdown-container {
+                .mat-form-field {
+                    display: block;
+                    margin-bottom: 15px;
+                }
+            }
+
+            .adf-site-dropdown-list-element {
+                width: 100%;
+                margin-bottom: 0;
+
+                .mat-select-trigger {
+                    font-size: 14px;
+                }
+
+            }
+
+            .adf-toolbar .mat-toolbar {
+                border-bottom-width: 0;
+                font-size: 14px;
+
+                &.mat-toolbar-single-row {
+                    height: auto;
+                }
+            }
+
+            &-breadcrumb {
+                .adf-dropdown-breadcumb-trigger {
+                    outline: none;
+                    .mat-icon {
+                        color: mat-color($foreground, base, 0.45);
+
+                        &:hover {
+                            color: mat-color($foreground, base, 0.65);
+                        }
+                    }
+                }
+
+                .adf-dropddown-breadcrumb-item-chevron {
+                    color: mat-color($foreground, base, 0.45);
+                }
+            }
+
+            &-list {
+                height: 200px;
+                overflow: auto;
+                border: 1px solid mat-color($foreground, base, 0.07);
+
+                .adf-highlight {
+                    color: mat-color($primary);
+                }
+
+                .adf-data-table {
+                    border: none;
+
+                    .adf-no-content-container {
+                        text-align: center;
+                    }
+
+                    thead {
+                        display: none;
+                    }
+
+                    .adf-data-table-cell {
+                        padding-top: 8px;
+                        padding-bottom: 8px;
+                        height: 30px;
+
+                        & .adf-name-location-cell-location {
+                            display: none;
+                        }
+
+                        & .adf-name-location-cell-name {
+                            padding: 0;
+                        }
+
+                        &--image {
+                            padding-left: 16px;
+                            padding-right: 8px;
+                        }
+
+                        &--text {
+                            padding-left: 8px;
+                        }
+                    }
+
+                    tbody tr {
+                        height: auto !important;
+
+                        &:first-child {
+                            .adf-data-table-cell {
+                                border-top: none;
+                            }
+                        }
+
+                        &:last-child {
+                            .adf-data-table-cell {
+                                border-bottom: none;
+                            }
+                        }
+                    }
+                }
+
+                &-searchLayout {
+
+                    .adf-data-table {
+                        .adf-data-table-cell {
+                            height: 56px;
+                            padding-bottom: 24px;
+
+                            & .adf-name-location-cell-location {
+                                display: block
+                            }
+
+                            & .adf-name-location-cell-name {
+                                padding: 18px 0 2px 0;
+                            }
+
+                            &.adf-content-selector-modified-cell {
+                                display: none;
+                            }
+
+                            &.adf-content-selector-modifier-cell {
+                                display: none;
+                            }
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+}
