+++ lib/content-services/content-node-selector/public-api.ts
@@ -16,5 +16,6 @@
  */
 
 export * from './content-node-selector.component-data.interface';
+export * from './content-node-selector-panel.component';
 export * from './content-node-selector.component';
 export * from './content-node-selector.service';
